demo/requirements.txt
locust
PyYAML
flask
